% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates_functions.R
\name{process_address}
\alias{process_address}
\title{Make client and billing parts in Rmd template}
\usage{
process_address(address, col_name = "Livraison")
}
\arguments{
\item{address}{list of address info (client or billing)}

\item{col_name}{header name of address}
}
\description{
Make client and billing parts in Rmd template
}
