% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates_functions.R
\name{process_bank}
\alias{process_bank}
\title{Process bank information for invoices}
\usage{
process_bank(bankinfo)
}
\arguments{
\item{bankinfo}{bank info list from YAML params}
}
\description{
Process bank information for invoices
}
