% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates_functions.R
\name{process_freelance}
\alias{process_freelance}
\title{Make freelance or company info suitable for HTML format}
\usage{
process_freelance(freelance)
}
\arguments{
\item{freelance}{list of freelance from the YAML params}
}
\description{
Make freelance or company info suitable for HTML format
}
