% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates_functions.R
\name{process_services}
\alias{process_services}
\title{Process services table}
\usage{
process_services(services, tva = "No")
}
\arguments{
\item{services}{services list from YAML parameterss}

\item{tva}{tva parameter}
}
\description{
Process services table
}
