% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates_functions.R
\name{process_total}
\alias{process_total}
\title{Process total table}
\usage{
process_total(services, tva = "No", discount = 0, deposit = 0)
}
\arguments{
\item{services}{services list from YAML params}

\item{tva}{"Yes"/"No"}

\item{discount}{discount rate (between 0 and 1)}

\item{deposit}{deposit amount. Only for bills}
}
\description{
Process total table
}
