% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_functions.R
\name{remove_header}
\alias{remove_header}
\title{Remove header of a HTML table}
\usage{
remove_header(x)
}
\arguments{
\item{x}{a HTML table with <thead> header tags}
}
\value{
the same HTML table but without the header part
}
\description{
Remove header of a HTML table
}
\examples{
library(knitr)
tab <- kable(iris, format = "html")
tab

#Removing header
remove_header(tab)

}
