% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_functions.R
\name{up_first}
\alias{up_first}
\title{Make first letter uppersase}
\usage{
up_first(st)
}
\arguments{
\item{st}{a string value}
}
\value{
The same string with the first letter as uppercase
}
\description{
Make first letter uppersase
}
\examples{
up_first("hello")
}
